Fill JWT with useful values
The Registered struct from the jwt crate contains some
fields that we will likely not need in the future. All
fields are marked with Option<T>, so we don't have to
provide them, but they still show up as null in the
resulting JWT which we may not want.

If so, we should provide our own struct for custom claims
that doesn't include these values.