Leverage Rocket features, Rust 2018, Clippy Lints
Leverages the databse feature of the Rocket framework and then passes
the connection from each request to the GraphQL context since it doesn't
need more than a single connection from the pool.

It also uses Rust 2018 where it can. This means using macros directly
from the crate instead of importing all of them via `#[macro_use]`.
Diesel doesn't allow this, AFAIK, though, so we fall back to the legacy
method of importing its macros.

Additionally, run `cargo clippy` to fix common lint issues. Not many
were found to begin with, which is a positive sign.

- ELD